@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Scraper\ScraperColissimo\Api;
 
@@ -8,17 +8,18 @@ use Scraper\ScraperColissimo\Entity\Response;
 
 final class ColissimoGenerateLabelApi extends ColissimoApi
 {
-    public function execute()
+    public function execute(): ColissimoGenerateLabel
     {
         $colissimoAdapter = ColissimoAdapter::execute($this->response);
-        $response         = $this->serializer->deserialize($colissimoAdapter->getJson(), Response::class, 'json');
+        /** @var Response $response */
+        $response = $this->serializer->deserialize($colissimoAdapter->getJson(), Response::class, 'json');
 
         $colissimoGenerateLabel = new ColissimoGenerateLabel();
-        $colissimoGenerateLabel->setResponse($response);
-        $colissimoGenerateLabel->setFile($colissimoAdapter->getFile());
+        $colissimoGenerateLabel->response = $response;
+        $colissimoGenerateLabel->file = $colissimoAdapter->getFile();
 
         if (!empty($colissimoAdapter->getCn23())) {
-            $colissimoGenerateLabel->setCn23($colissimoAdapter->getCn23());
+            $colissimoGenerateLabel->cn23 = $colissimoAdapter->getCn23();
         }
 
         return $colissimoGenerateLabel;
