@@ -1,48 +1,10 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Scraper\ScraperColissimo\Entity;
 
 final class Message
 {
-    private string $id;
-
-    private string $messageContent;
-
-    private string $type;
-
-    public function getId(): string
-    {
-        return $this->id;
-    }
-
-    public function setId(string $id): self
-    {
-        $this->id = $id;
-
-        return $this;
-    }
-
-    public function getMessageContent(): string
-    {
-        return $this->messageContent;
-    }
-
-    public function setMessageContent(string $messageContent): self
-    {
-        $this->messageContent = $messageContent;
-
-        return $this;
-    }
-
-    public function getType(): string
-    {
-        return $this->type;
-    }
-
-    public function setType(string $type): self
-    {
-        $this->type = $type;
-
-        return $this;
-    }
+    public string $id;
+    public string $messageContent;
+    public string $type;
 }
