@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Scraper\ScraperColissimo\Rest;
 
@@ -17,7 +17,7 @@ final class GenerateLabelRequest
     public function __construct()
     {
         $this->outputFormat = new OutputFormat();
-        $this->letter       = new Letter();
+        $this->letter = new Letter();
     }
 
     public function getOutputFormat(): OutputFormat
