@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Scraper\ScraperColissimo\Rest;
 
@@ -16,9 +16,9 @@ final class Letter
 
     public function __construct()
     {
-        $this->service   = new Service();
-        $this->parcel    = new Parcel();
-        $this->sender    = new Sender();
+        $this->service = new Service();
+        $this->parcel = new Parcel();
+        $this->sender = new Sender();
         $this->addressee = new Addressee();
     }
 
@@ -27,7 +27,7 @@ final class Letter
         return $this->customsDeclarations;
     }
 
-    public function setCustomsDeclarations(?CustomsDeclarations $customsDeclarations = null): self
+    public function setCustomsDeclarations(CustomsDeclarations $customsDeclarations = null): self
     {
         $this->customsDeclarations = $customsDeclarations;
 
