@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Scraper\ScraperColissimo\Rest;
 
@@ -101,7 +101,7 @@ final class Parcel
         return $this->instructions;
     }
 
-    public function setInstructions(?string $instructions = null): self
+    public function setInstructions(string $instructions = null): self
     {
         $this->instructions = $instructions;
 
@@ -113,7 +113,7 @@ final class Parcel
         return $this->pickupLocationId;
     }
 
-    public function setPickupLocationId(?string $pickupLocationId = null): self
+    public function setPickupLocationId(string $pickupLocationId = null): self
     {
         $this->pickupLocationId = $pickupLocationId;
 
