@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Scraper\ScraperColissimo\Rest;
 
@@ -13,7 +13,7 @@ final class Sender
         $this->address = new Address();
     }
 
-    public function setSenderParcelRef(?string $senderParcelRef = null): self
+    public function setSenderParcelRef(string $senderParcelRef = null): self
     {
         $this->senderParcelRef = $senderParcelRef;
 
